Refactor UI elements
