+++ battle_map_tv/elements/sliders.py
@@ -1,7 +1,7 @@
-from PySide6.QtWidgets import QWidget, QVBoxLayout, QLabel, QHBoxLayout, QGridLayout
-from PySide6.QtCore import Signal
+from PySide6.QtWidgets import QWidget, QVBoxLayout, QLabel, QHBoxLayout, QGridLayout, QSlider
+from PySide6.QtCore import Signal, Qt
 
-from battle_map_tv.ui_elements import StyledSlider, StyledLineEdit
+from .text_based import StyledLineEdit
 
 
 class DualScaleSlider(QWidget):
@@ -76,3 +76,30 @@ class DualScaleSlider(QWidget):
 
     def _value_bounds(self, value: float) -> float:
         return min(max(value, 1 / self.factor_fine), self.max_scale)
+
+
+class StyledSlider(QSlider):
+    def __init__(self, lower: int, upper: int, default: int, *args, **kwargs):
+        super().__init__(Qt.Horizontal, *args, **kwargs)  # type: ignore[attr-defined]
+        self.setMinimum(lower)
+        self.setMaximum(upper)
+        self.setValue(default)
+        self.setStyleSheet(
+            """
+            QSlider {
+                height: 40px;
+            }
+            QSlider::groove:horizontal {
+                height: 10px;
+                background: #404040;
+                margin: 0px;
+            }
+            QSlider::handle:horizontal {
+                background: #717173;
+                border: 1px solid #3E3E40;
+                width: 20px;
+                margin: -15px 0;
+                border-radius: 6px;
+            }
+        """
+        )
